[module4-task1] Service/CLI/Service: rename services
