+++ src/service/cli/server/routes/offers.ts
@@ -1,5 +1,5 @@
 import {Router, Request, Response} from "express";
-import OffersService from "../services/offers-service";
+import OffersService from "../services/offers.service";
 import {HttpCodes} from "../../../../shared/http-codes";
 
 const offersRouter: Router = Router();
